move board declaration to GameObject
